drf/[[ cookiecutter.project_slug ]]/requirements.txt
click==8.1.8
cryptography==44.0.2
Django==5.2
django-jazzmin==3.0.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
gunicorn==23.0.0
l4py==0.1.9
royman-dotenv==1.1.2
whitenoise==6.9.0
